requirements.txt
# huggingface datasets
datasets
# gemini
google-genai
# general
numpy==1.26.4
pandas